+++ src/main/java/gift/controller/dto/response/ProductResponse.java
@@ -1,4 +1,4 @@
-package gift.controller.res;
+package gift.controller.dto.response;
 
 import gift.model.Product;
